// src/types/getAppointmentsParams.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Hopital Management System Api
 * OpenAPI spec version: v1
 */

export type GetAppointmentsParams = {
    'PatientId.Equal'?: number;
    'PatientId.NotEquals'?: number;
    'PatientId.GreaterThan'?: number;
    'PatientId.GreaterThanOrEqual'?: number;
    'PatientId.LessThan'?: number;
    'PatientId.LessThanOrEqual'?: number;
    'PatientId.In'?: number[];
    'PatientId.IsNull'?: boolean;
    'PatientId.IsNotNull'?: boolean;
    'DoctorId.Equal'?: number;
    'DoctorId.NotEquals'?: number;
    'DoctorId.GreaterThan'?: number;
    'DoctorId.GreaterThanOrEqual'?: number;
    'DoctorId.LessThan'?: number;
    'DoctorId.LessThanOrEqual'?: number;
    'DoctorId.In'?: number[];
    'DoctorId.IsNull'?: boolean;
    'DoctorId.IsNotNull'?: boolean;
    'AppointmentDate.Equal'?: string;
    'AppointmentDate.NotEquals'?: string;
    'AppointmentDate.GreaterThan'?: string;
    'AppointmentDate.GreaterThanOrEqual'?: string;
    'AppointmentDate.LessThan'?: string;
    'AppointmentDate.LessThanOrEqual'?: string;
    'AppointmentDate.IsNull'?: boolean;
    'AppointmentDate.IsNotNull'?: boolean;
    'Status.Contains'?: string;
    'Status.NotContains'?: string;
    'Status.Equal'?: string;
    'Status.NotEquals'?: string;
    'Status.StartsWith'?: string;
    'Status.EndsWith'?: string;
    'Status.IsNull'?: boolean;
    'Status.IsNotNull'?: boolean;
    'Status.In'?: string[];
    'CreatedAt.Equal'?: string;
    'CreatedAt.NotEquals'?: string;
    'CreatedAt.GreaterThan'?: string;
    'CreatedAt.GreaterThanOrEqual'?: string;
    'CreatedAt.LessThan'?: string;
    'CreatedAt.LessThanOrEqual'?: string;
    'CreatedAt.IsNull'?: boolean;
    'CreatedAt.IsNotNull'?: boolean;
    'AppointmentTime.Equal'?: number;
    'AppointmentTime.NotEquals'?: number;
    'AppointmentTime.GreaterThan'?: number;
    'AppointmentTime.GreaterThanOrEqual'?: number;
    'AppointmentTime.LessThan'?: number;
    'AppointmentTime.LessThanOrEqual'?: number;
    'AppointmentTime.In'?: number[];
    'AppointmentTime.IsNull'?: boolean;
    'AppointmentTime.IsNotNull'?: boolean;
    Search?: string;
    Page?: number;
    PageSize?: number;
    Skip?: number;
    Take?: number;
    SortBy?: string;
    SortOrder?: string;
};